src/test/javascript/spec/app/entities/signos-vitales/signos-vitales.service.spec.ts
import { TestBed, getTestBed } from '@angular/core/testing';
import { HttpClientTestingModule, HttpTestingController } from '@angular/common/http/testing';
import { SignosVitalesService } from 'app/entities/signos-vitales/signos-vitales.service';
import { ISignosVitales, SignosVitales } from 'app/shared/model/signos-vitales.model';

describe('Service Tests', () => {
  describe('SignosVitales Service', () => {
    let injector: TestBed;
    let service: SignosVitalesService;
    let httpMock: HttpTestingController;
    let elemDefault: ISignosVitales;
    let expectedResult: ISignosVitales | ISignosVitales[] | boolean | null;

    beforeEach(() => {
      TestBed.configureTestingModule({
        imports: [HttpClientTestingModule]
      });
      expectedResult = null;
      injector = getTestBed();
      service = injector.get(SignosVitalesService);
      httpMock = injector.get(HttpTestingController);

      elemDefault = new SignosVitales(0, 0, 0, 0, 0);
    });

    describe('Service methods', () => {
      it('should find an element', () => {
        const returnedFromService = Object.assign({}, elemDefault);

        service.find(123).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'GET' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(elemDefault);
      });

      it('should create a SignosVitales', () => {
        const returnedFromService = Object.assign(
          {
            id: 0
          },
          elemDefault
        );

        const expected = Object.assign({}, returnedFromService);

        service.create(new SignosVitales()).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'POST' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(expected);
      });

      it('should update a SignosVitales', () => {
        const returnedFromService = Object.assign(
          {
            presionArterial: 1,
            respiracion: 1,
            pulso: 1,
            temperatura: 1
          },
          elemDefault
        );

        const expected = Object.assign({}, returnedFromService);

        service.update(expected).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'PUT' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(expected);
      });

      it('should return a list of SignosVitales', () => {
        const returnedFromService = Object.assign(
          {
            presionArterial: 1,
            respiracion: 1,
            pulso: 1,
            temperatura: 1
          },
          elemDefault
        );

        const expected = Object.assign({}, returnedFromService);

        service.query().subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'GET' });
        req.flush([returnedFromService]);
        httpMock.verify();
        expect(expectedResult).toContainEqual(expected);
      });

      it('should delete a SignosVitales', () => {
        service.delete(123).subscribe(resp => (expectedResult = resp.ok));

        const req = httpMock.expectOne({ method: 'DELETE' });
        req.flush({ status: 200 });
        expect(expectedResult);
      });
    });

    afterEach(() => {
      httpMock.verify();
    });
  });
});
